Abstract the cobra comand build to a separate package
